feat(store): user & event modules created
