+++ src/store/index.js
@@ -1,15 +1,12 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
-import EventService from '@/services/EventService'
+
+import { user, event } from '@/store/modules'
 
 Vue.use(Vuex)
 
 export default new Vuex.Store({
   state: {
-    user: { id: 'abc123', name: 'David' },
-    event: {},
-    events: [],
-    eventsTotal: 0,
     categories: [
       'sustainability',
       'nature',
@@ -20,45 +17,8 @@ export default new Vuex.Store({
       'community',
     ],
   },
-  getters: {
-    getEventById: (state) => (id) => state.events.filter((e) => (e.id = id))[0],
+  modules: {
+    user,
+    event,
   },
-  mutations: {
-    ADD_EVENT(state, event) {
-      state.events.push(event)
-    },
-    SET_EVENT(state, event) {
-      state.event = event
-    },
-    SET_EVENTS(state, events) {
-      state.events = events
-    },
-    SET_EVENTS_TOTAL(state, eventsTotal) {
-      state.eventsTotal = eventsTotal
-    },
-  },
-  actions: {
-    createEvent({ commit }, event) {
-      return EventService.addEvent(event).then(() => commit('ADD_EVENT', event))
-    },
-    fetchEvent({ commit, getters }, id) {
-      const event = getters.getEventById(id)
-      if (event) {
-        commit('SET_EVENT', event)
-      } else {
-        EventService.getEvent(id)
-          .then((response) => commit('SET_EVENT', response.data))
-          .catch((error) => console.error(error.response))
-      }
-    },
-    fetchEvents({ commit }, { perPage, page }) {
-      EventService.getEvents(perPage, page)
-        .then((response) => {
-          commit('SET_EVENTS_TOTAL', response.headers['x-total-count'])
-          commit('SET_EVENTS', response.data)
-        })
-        .catch((error) => console.error(error.response))
-    },
-  },
-  modules: {},
 })
