+++ src/store/modules/index.js
@@ -1,0 +1,2 @@
+export * as user from './user'
+export * as event from './event'
