+++ src/views/EventCreate.vue
@@ -72,13 +72,13 @@ export default {
   },
   methods: {
     createFreshEvent() {
-      const user = this.$store.state.user
+      const user = this.$store.state.user.user
       const id = Math.floor(Math.random() * 10000000)
 
       return {
         id,
         category: '',
-        user: user,
+        user,
         organizer: user.name,
         title: '',
         description: '',
