+++ src/views/EventList.vue
@@ -1,7 +1,7 @@
 <template>
   <div>
-    <h1>Event listing</h1>
-    <EventCard v-for="event in events" :key="event.id" :event="event" />
+    <h1>Events for {{ user.user.name }}</h1>
+    <EventCard v-for="event in event.events" :key="event.id" :event="event" />
     <template v-if="page !== 1">
       <router-link
         :to="{ name: 'event-list', query: { page: page - 1 } }"
@@ -11,7 +11,7 @@
       </router-link>
     </template>
 
-    <template v-if="eventsTotal > page * perPage">
+    <template v-if="hasNextPage">
       <router-link
         :to="{ name: 'event-list', query: { page: page + 1 } }"
         rel="next"
@@ -36,15 +36,18 @@ export default {
     }
   },
   created() {
-    this.$store.dispatch('fetchEvents', {
+    this.$store.event.dispatch('fetchEvents', {
       perPage: this.perPage,
       page: this.page,
     })
   },
   computed: {
-    ...mapState(['events', 'eventsTotal']),
+    ...mapState(['event', 'user']),
     page() {
       return parseInt(this.$route.query.page) || 1
+    },
+    hasNextPage() {
+      return this.event.eventsTotal > this.page * this.perPage
     },
   },
 }
