+++ src/views/EventShow.vue
@@ -33,7 +33,9 @@ export default {
   created() {
     this.$store.dispatch('fetchEvent', this.id)
   },
-  computed: mapState(['event']),
+  computed: mapState({
+    event: (state) => state.event.event,
+  }),
 }
 </script>
